changed views to apps
