+++ src/apps/SlimApp.jsx
